Workaround: avoid pushing updated tags
In some workflows, the CI is triggered when a tag is pushed, even if the
tag is being updated and linked to a new SHA.

If the CI is only looking at the code, it is not needed to re-launch a
new job. For that, the use_git_push_option can be used and a "ci.skip"
option will be added when a tag is updated.

But some CI instances ignores "ci.skip" git push option. In this case,
it makes not to push the new tag that will trigger a new CI job
re-validating what was already done before.