tags: option not to update series on tags
Currently, when a new tag is sent on the mailing list, the series is
updated in the Git repository.

The push is done with '-o ci.skip', but this seems to be something specific to
GitLab. It has no effects with GitHub for example. Not to waste resources on
retesting everything and sending notifications once a tag is shared, a new
option has been added:

  [default]
  update_series_on_tags = true

Not to change the current behaviour, the series will be updated on new tags by
default. The option can be set to 'false' (or: 'n', 'no', 'f', 'off', '0') to
disable this feature.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>